added mentioned role and added embed category
got the category in jolly bean code (he gived me perm)
