+++ src/blocks/discord/embed/create_embed_then.js
@@ -14,7 +14,7 @@ const blockData = {
             "type":"input_dummy"
         }
     ],
-    "colour": "#4C97FF",
+    "colour": "#40BF4A",
     "previousStatement": null,
     "nextStatement": null,
     "tooltip": "",
