+++ src/blocks/discord/embed/create_embed_then_set_author.js
@@ -1,18 +1,23 @@
 import * as Blockly from "blockly/core";
 import { registerRestrictions } from "../../../restrictions";
 
-const blockName = "s4d_create_embed_then_set_thumbnail";
+const blockName = "s4d_create_embed_then_set_author";
 
 const blockData = {
-    "message0": "%{BKY_CREATE_EMBED_THEN_SET_THUMBNAIL}",
+    "message0": "%{BKY_CREATE_EMBED_THEN_SET_AUTHOR}",
     "args0": [
         {
             "type":"input_value",
-            "name":"THUMBNAIL",
+            "name":"AUTHOR",
             "check":"String"
+        },
+        {
+            "type":"input_value",
+            "name":"PFP",
+            "check":"String"  
         }
     ],
-    "colour": "#4C97FF",
+    "colour": "#40BF4A",
     "previousStatement": null,
     "nextStatement": null,
     "inputsInline": true,
@@ -27,8 +32,9 @@ Blockly.Blocks[blockName] = {
 };
 
 Blockly.JavaScript[blockName] = function(block){
-    const thumbnail = Blockly.JavaScript.valueToCode(block, "THUMBNAIL", Blockly.JavaScript.ORDER_ATOMIC);
-    const code = `embed.setThumbnail(${thumbnail});\n`;
+    const author = Blockly.JavaScript.valueToCode(block, "AUTHOR", Blockly.JavaScript.ORDER_ATOMIC);
+    const pfp = Blockly.JavaScript.valueToCode(block, "PFP", Blockly.JavaScript.ORDER_ATOMIC);
+    const code = `embed.setAuthor(${author},${pfp});\n`;
     return code;
 };
