+++ src/blocks/discord/embed/create_embed_then_set_description.js
@@ -1,15 +1,20 @@
 import * as Blockly from "blockly/core";
 import { registerRestrictions } from "../../../restrictions";
 
-const blockName = "s4d_create_embed_then_send";
+const blockName = "s4d_create_embed_then_set_description";
 
 const blockData = {
-    "message0": "%{BKY_CREATE_EMBED_THEN_SEND}",
+    "message0": "%{BKY_CREATE_EMBED_THEN_SET_DESCRIPTION}",
     "args0": [
+        {
+            "type":"input_value",
+            "name":"DESCRIPTION",
+            "check":"String"
+        }
     ],
-    "colour": "#4C97FF",
+    "colour": "#40BF4A",
     "previousStatement": null,
-    "nextStatement": null,    
+    "nextStatement": null,
     "inputsInline": true,
     "tooltip": "",
     "helpUrl": ""
@@ -21,8 +26,9 @@ Blockly.Blocks[blockName] = {
     }
 };
 
-Blockly.JavaScript[blockName] = function(){
-    const code = `s4dmessage.channel.send(embed);\n`;
+Blockly.JavaScript[blockName] = function(block){
+    const description = Blockly.JavaScript.valueToCode(block, "DESCRIPTION", Blockly.JavaScript.ORDER_ATOMIC);
+    const code = `embed.setDescription(${description});\n`;
     return code;
 };
