+++ src/blocks/discord/embed/create_embed_then_set_footer.js
@@ -1,18 +1,23 @@
 import * as Blockly from "blockly/core";
 import { registerRestrictions } from "../../../restrictions";
 
-const blockName = "s4d_create_embed_then_set_color";
+const blockName = "s4d_create_embed_then_set_footer";
 
 const blockData = {
-    "message0": "%{BKY_CREATE_EMBED_THEN_SET_COLOR}",
+    "message0": "%{BKY_CREATE_EMBED_THEN_SET_FOOTER}",
     "args0": [
         {
             "type":"input_value",
-            "name":"COLOR",
-            "check":"Colour"
+            "name":"FOOTER",
+            "check":"String"
+        },
+        {
+            "type":"input_value",
+            "name":"IMAGE",
+            "check":"String"
         }
     ],
-    "colour": "#4C97FF",
+    "colour": "#40BF4A",
     "previousStatement": null,
     "nextStatement": null,
     "inputsInline": true,
@@ -27,8 +32,9 @@ Blockly.Blocks[blockName] = {
 };
 
 Blockly.JavaScript[blockName] = function(block){
-    const color = Blockly.JavaScript.valueToCode(block, "COLOR", Blockly.JavaScript.ORDER_ATOMIC);
-    const code = `embed.setColor(${color});\n`;
+    const footer = Blockly.JavaScript.valueToCode(block, "FOOTER", Blockly.JavaScript.ORDER_ATOMIC);
+    const footerimage = Blockly.JavaScript.valueToCode(block, "IMAGE", Blockly.JavaScript.ORDER_ATOMIC);
+    const code = `embed.setFooter(${footer},${footerimage});\n`;
     return code;
 };
