+++ src/blocks/discord/embed/create_embed_then_set_title.js
@@ -1,23 +1,18 @@
 import * as Blockly from "blockly/core";
 import { registerRestrictions } from "../../../restrictions";
 
-const blockName = "s4d_create_embed_then_set_footer";
+const blockName = "s4d_create_embed_then_set_title";
 
 const blockData = {
-    "message0": "%{BKY_CREATE_EMBED_THEN_SET_FOOTER}",
+    "message0": "%{BKY_CREATE_EMBED_THEN_SET_TITLE}",
     "args0": [
         {
             "type":"input_value",
-            "name":"FOOTER",
-            "check":"String"
-        },
-        {
-            "type":"input_value",
-            "name":"IMAGE",
+            "name":"TITLE",
             "check":"String"
         }
     ],
-    "colour": "#4C97FF",
+    "colour": "#40BF4A",
     "previousStatement": null,
     "nextStatement": null,
     "inputsInline": true,
@@ -32,9 +27,8 @@ Blockly.Blocks[blockName] = {
 };
 
 Blockly.JavaScript[blockName] = function(block){
-    const footer = Blockly.JavaScript.valueToCode(block, "FOOTER", Blockly.JavaScript.ORDER_ATOMIC);
-    const footerimage = Blockly.JavaScript.valueToCode(block, "IMAGE", Blockly.JavaScript.ORDER_ATOMIC);
-    const code = `embed.setFooter(${footer},${footerimage});\n`;
+    const title = Blockly.JavaScript.valueToCode(block, "TITLE", Blockly.JavaScript.ORDER_ATOMIC);
+    const code = `embed.setTitle(${title});\n`;
     return code;
 };
