+++ src/blocks/discord/embed/index.js
@@ -1,0 +1,10 @@
+import "./create_embed_then_send";
+import "./create_embed_then_set_color";
+import "./create_embed_then_set_title";
+import "./create_embed_then_set_author";
+import "./create_embed_then_set_description";
+import "./create_embed_then_set_thumbnail";
+import "./create_embed_then_add_field";
+import "./create_embed_then_set_image";
+import "./create_embed_then_set_footer";
+import "./create_embed_then";
