+++ src/blocks/discord/message-event/index.js
@@ -18,14 +18,5 @@ import "./send_reply_then_edit";
 import "./send_reply_then_delete";
 import "./send_reply_then_add_reaction";
 import "./send_reply_then_reply_id";
-import "./create_embed_then";
 import "./mentioned_channel";
-import "./create_embed_then_send";
-import "./create_embed_then_set_color";
-import "./create_embed_then_set_title";
-import "./create_embed_then_set_author";
-import "./create_embed_then_set_description";
-import "./create_embed_then_set_thumbnail";
-import "./create_embed_then_add_field";
-import "./create_embed_then_set_image";
-import "./create_embed_then_set_footer";
+import "./mentioned_role";
