+++ src/blocks/discord/message-event/mentioned_role.js
@@ -1,0 +1,33 @@
+import Blockly from "blockly/core";
+import { registerRestrictions } from "../../../restrictions";
+
+const blockName = "s4d_mentioned_role";
+
+const blockData = {
+    "message0": "%{BKY_MENTIONED_ROLE}",
+    "colour": "#187795",
+    "tooltip": "",
+    "output": "Role",
+    "helpUrl": ""
+};
+
+Blockly.Blocks[blockName] = {
+    init: function() {
+        this.jsonInit(blockData);
+    }
+};
+
+Blockly.JavaScript[blockName] = function() {
+    const code = [`s4dmessage.mentions.roles.first()`, Blockly.JavaScript.ORDER_NONE];
+    return code;
+};
+
+registerRestrictions(blockName, [
+    {
+        type: "toplevelparent",
+        message: "RES_MUST_BE_IN_ON_MESSAGE",
+        types: [
+            "s4d_on_message"
+        ]
+    }
+]);
