+++ src/locales/en-US.json
@@ -161,6 +161,7 @@
   "OTHERS":"other",
   "NEWLINE": "new line",
   "CREATE_EMBED_THEN":"create new embed %2 then %1",
+  "SEND_EMBED": "Send the embed",
   "MENTIONED_CHANNEL":"mentioned channel",
   "RES_SET_SERVER_NAME_NEW_NICKNAME": "You must specify a new valid name for the server!",
   "REACTION_WAIT_ADDED_REACTION":"wait for reaction of member %1 during (minutes) %2 when reaction is added %3",
@@ -245,7 +246,12 @@
   "TOOLBOX_E_DELETED":"deleted",
   "ON_DELETED":"when a message is deleted",
   "GET_WEBHOOK_THEN":"get webhook with id %1 and token %2 then %3 %4",
+  "COLOR_MESSAGE": "type RANDOM in text block instead of (random color block)",
   "GET_ALL_ROLE_ROLE":"role",
+  "CATEGORY_E_EMBED":"embed",
+  "HYPERLINK_MESSAGE": "To add hyperlink, use format : [name](link)",
+  "HYPERLINK_MESSAGE_2": "hyperlinks can be only used in descriptions",
+  "MENTIONED_ROLE":"mentioned role",
   "SEND_WEBHOOK":"send webhook %1",
   "RES_GET_WEBHOOK_PARENT":"this block can only be used in a get webhook block!",
   "GET_ALL_ROLE":"get all roles in server %1 then %2 %3",
