+++ src/locales/pt-BR.json
@@ -144,7 +144,13 @@
   "GET_WEBHOOK_THEN":"pegue webhook com o id %1 e o token %2 ai %3 %4",
   "RES_GET_WEBHOOK_PARENT":"este bloco so pode ser usado em um pegue webhook!",
   "SEND_WEBHOOK":"envie webhook %1",
+  "MENTIONED_ROLE":"cargo mencionado",
+  "CATEGORY_E_EMBED":"embed",
+  "HYPERLINK_MESSAGE": "para adicionar hyperlink, use o formato : [nome](link)",
+  "HYPERLINK_MESSAGE_2": "hyperlinks so pode ser usado na descrição",
+  "COLOR_MESSAGE": "escreva RANDOM no texto ao invez de (random color block)",
   "GET_ALL_SERVER_SERVER":"servidor",
+  "SEND_EMBED": "Envie o embed",
   "GET_ALL_CHANNEL":"pega todos os canais no servido %1 ai %2 %3",
   "WAIT_REACT_EMOJI_ID":"id do emoji",
   "GET_ALL_CHANNEL_CHANNEL":"canal",
